# specs/003-pkm-refactor/contracts/openapi.yaml
openapi: '3.0.3'
info:
  title: PKD — Personal Knowledge Management API
  version: '0.2.0'
  description: |
    API for the PKM refactor (003-pkm-refactor).
    Extends the 001 API with bidirectional links, graph data, and content capture.
servers:
  - url: /
security:
  - sessionCookie: []
    csrfHeader: []
paths:
...
  /api/documents/{id}/shares:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    post:
      summary: Create share link
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  url: { type: string }
                  revoke_id: { type: integer }
  /api/documents/{id}/shares/{shareId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
      - name: shareId
        in: path
        required: true
        schema: { type: integer }
    delete:
      summary: Revoke share link
      responses:
        '204': { description: Revoked }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pkd_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    DocumentId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...